requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
langchain
langchain-openai
pydantic
openai
typing-extensions
python-multipart
pillow
tavily-python